feature(ingest): make superdesk.io.registry a resource with get endpoints
