+++ superdesk/io/feed_parsers/afp_newsml_1_2.py
@@ -9,7 +9,7 @@
 # at https://www.sourcefabric.org/superdesk/license
 
 from superdesk.io.feed_parsers.newsml_1_2 import NewsMLOneFeedParser
-from superdesk.io import register_feed_parser
+from superdesk.io.registry import register_feed_parser
 from superdesk.utc import utcnow
 from pytz import utc
