+++ superdesk/io/feed_parsers/dpa_iptc7901.py
@@ -10,7 +10,7 @@
 # at https://www.sourcefabric.org/superdesk/license
 
 from .iptc7901 import IPTC7901FeedParser
-from superdesk.io import register_feed_parser
+from superdesk.io.registry import register_feed_parser
 from flask import current_app as app
 from superdesk.metadata.item import ITEM_TYPE, CONTENT_TYPE
