+++ superdesk/io/feed_parsers/newsml_1_2.py
@@ -10,7 +10,7 @@
 
 import datetime
 from superdesk.etree import etree
-from superdesk.io import register_feed_parser
+from superdesk.io.registry import register_feed_parser
 from superdesk.io.feed_parsers import XMLFeedParser
 from superdesk.io.iptc import subject_codes
 from superdesk.errors import ParserError
