+++ superdesk/io/feed_parsers/pa_nitf.py
@@ -10,7 +10,7 @@
 # at https://www.sourcefabric.org/superdesk/license
 
 from superdesk.io.feed_parsers.nitf import NITFFeedParser, SkipValue
-from superdesk.io import register_feed_parser
+from superdesk.io.registry import register_feed_parser
 import re
 from bs4 import BeautifulSoup
 import xml.etree.ElementTree as etree
