--- superdesk/io/register.py
@@ -1,8 +1,0 @@
-
-registered_feed_parsers = {}
-allowed_feed_parsers = []
-
-registered_feeding_services = {}
-allowed_feeding_services = []
-feeding_service_errors = {}
-publish_errors = []
